Moved scripts to start setting up push of local objects
