+++ Assets/Scripts/Environment/Pushable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0aaacf71275dc734182f20b3145839b2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
