+++ Assets/Scripts/Environment/Pushable/IPushable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3d3c8e31485313a47a6179a16d2aa0cf
+guid: 803ced2f538960d4594103c857f85775
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
