+++ Assets/Scripts/Environment/Pushable/Pushable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3b65b70a1a4a9404abccddc6c944a40e
+guid: a50731e36aed2d446bfbd69078224021
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
